move bubbletea files outside of cmd package
